+++ bot.js
@@ -3,7 +3,8 @@ require("dotenv").config();
 const Discord = require("discord.js");
 const Client = require("./utils/Client");
 const chalk = require("chalk");
-const { errorLog, fileLoader } = require("./coreFunctions.js");
+const { errorLog } = require("./utils/logs");
+const { fileLoader } = require("./utils/misc.js");
 const { connect, connection } = require("mongoose");
 const autoIncrement = require("mongoose-sequence");
 const { basename } = require("path");
