+++ commands/server admin/autosetup.js
@@ -1,5 +1,5 @@
 const { emoji } = require("../../config.json");
-const { dbQuery, dbModify } = require("../../coreFunctions.js");
+const { dbQuery, dbModify } = require("../../utils/db");
 const { string } = require("../../utils/strings");
 module.exports = {
 	controls: {
