+++ commands/server admin/bean.js
@@ -1,4 +1,5 @@
-const { dbQuery, fetchUser } = require("../../coreFunctions");
+const { fetchUser } = require("../../utils/misc");
+const { dbQuery } = require("../../utils/db");
 const { string } = require("../../utils/strings");
 module.exports = {
 	controls: {
