--- events/disconnect.js
@@ -1,5 +1,0 @@
-const { coreLog } = require("../coreFunctions.js");
-module.exports = (Discord, client, event) => {
-	console.log(event);
-	coreLog(`:rotating_light: **Websocket Disconnect**\n>>> Code: ${event.code}\nReason: ${event.reason}`);
-};
