--- events/resume.js
@@ -1,4 +1,0 @@
-const { coreLog } = require("../coreFunctions.js");
-module.exports = (Discord, client) => {
-	coreLog("🆙 **Websocket Resumed**", client);
-};
