+++ index.js
@@ -5,7 +5,6 @@ const { ShardingManager } = require("discord.js");
 
 const manager = new ShardingManager("./bot.js", { token: process.env.TOKEN });
 
-
 manager.spawn();
 
 manager.on("shardCreate", (shard) => {
